(trunk) Potential fix for g++ 4.1.
git-svn-id: http://svn.gplates.org/projects/gplates/trunk@8175 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/gui/TopologyTools.cc
@@ -898,7 +898,7 @@ GPlatesGui::TopologyTools::can_insert_focused_feature_into_topology() const
 	// Make sure that we don't create a contiguous sequence of three sections with the
 	// same geometry when we insert the focused feature geometry.
 	if (std::find(topology_sections_indices.begin(), topology_sections_indices.end(),
-			get_prev_index(insert_section_index, d_section_info_seq))
+			static_cast<int>(get_prev_index(insert_section_index, d_section_info_seq)))
 				!= topology_sections_indices.end())
 	{
 		// Cannot insert same geometry after itself.
